+++ config/index.js
@@ -18,6 +18,11 @@ const defaultRoot = resolve(process.env.HOME, ".cynomys", "protal");
 
 const [num, dataDir, logDir] = [cpus().length, resolve(defaultRoot, "data"), resolve(defaultRoot, "logs")];
 
+siteConfig.set("remote", { // 远程节点设置
+  enable : true, // 是否开启远程
+  host: "namo", // 远程服务器ip
+  port: 15024 // 远程服务对外节点
+});
 siteConfig.set("tmpDir", "/tmp"); // 临时目录，用于存放上传的临时文件
 siteConfig.set("staticDir", "/tmp/staticDir"); // 静态资源目录，用于存放静态资源
 siteConfig.set("sourceHost", "http://almcloud.jwis.cn/"); // Nginx 静态资源地址